build warning cleaning

+++ groups/btl/btls/btls_reservationguard.t.cpp
@@ -346,7 +346,7 @@ int main(int argc, char *argv[]) {
                                              bucket.calculateTimeToSubmit(now);
             bsls::Types::Uint64 uS = timeToSubmit.totalMicroseconds() +
                                    (timeToSubmit.nanoseconds() % 1000) ? 1 : 0;
-            bdlqq::ThreadUtil::microSleep(uS);
+            bdlqq::ThreadUtil::microSleep(static_cast<int>(uS));
         }
     }
 //..
